Edit game reward, ui

+++ Assets/02.Scripts/PlayerPanelStatus.cs
@@ -16,11 +16,11 @@ public class PlayerPanelStatus : MonoBehaviour
 
     void Update()
     {
-        panelStatusTMP[0].text = "배고픔 : " + StatusBar.instance.curHunger + "%";
-        panelStatusTMP[1].text = "깨끗함 : " + StatusBar.instance.curClean + "%";
-        panelStatusTMP[2].text = "똑똑함 : " + StatusBar.instance.curSmart + "%";
-        panelStatusTMP[3].text = "움직임 : " + StatusBar.instance.curActive + "%";
-        panelStatusTMP[4].text = "기력 : " + StatusBar.instance.curEnergy + "%";
-        panelStatusTMP[5].text = "행복 : " + StatusBar.instance.curHappy + "%";
+        panelStatusTMP[0].text = "배고픔 : " + (int)StatusBar.instance.curHunger + "%";
+        panelStatusTMP[1].text = "깨끗함 : " + (int)StatusBar.instance.curClean + "%";
+        panelStatusTMP[2].text = "똑똑함 : " + (int)StatusBar.instance.curSmart + "%";
+        panelStatusTMP[3].text = "움직임 : " + (int)StatusBar.instance.curActive + "%";
+        panelStatusTMP[4].text = "기력 : " + (int)StatusBar.instance.curEnergy + "%";
+        panelStatusTMP[5].text = "행복 : " + (int)StatusBar.instance.curHappy + "%";
     }
 }
